refactor: complete plural naming consistency in backend/common folders
- Rename backend/common/domain → domains
- Rename backend/common/domain/repository → domains/repositories
- Rename backend/common/infrastructure → infrastructures
- Update all import paths in API routes and business logic
- Update coding convention documentation with plural naming rules
- Update REFACTOR_STATUS.md with completion status

All builds and lints now pass with consistent plural folder naming throughout the codebase.
