+++ app/api/admin/reservations/(adaptor)/route.ts
@@ -1,8 +1,8 @@
-import { SbRsvRepository } from '@/backend/common/infrastructure/repositories/SbRsvRepository';
+import { SbRsvRepository } from '@/backend/common/infrastructures/repositories/SbRsvRepository';
 import { NextRequest, NextResponse } from 'next/server';
 import { GetRsvListUsecase } from '@/backend/admin/reservations/usecases/GetRsvListUsecase';
 import { SupabaseClient } from '@supabase/supabase-js';
-import { createClient } from '@/backend/common/infrastructure/supabase/server';
+import { createClient } from '@/backend/common/infrastructures/supabase/server';
 
 export async function GET(request: NextRequest) {
   const token = request.headers.get('Authorization');
