+++ app/api/auth/signup/(adaptor)/route.ts
@@ -1,12 +1,12 @@
 import { NextRequest, NextResponse } from 'next/server';
 import { RegisterUseCase } from '@/backend/auth/signup/usecases/RegisterUseCase'; 
-import { SupabaseUserRepository } from '@/backend/common/infrastructure/repositories/SbUserRepository';
+import { SupabaseUserRepository } from '@/backend/common/infrastructures/repositories/SbUserRepository';
 import { SignupResponse, SignupErrorResponse } from '@/backend/auth/signup/dtos/SignupResponse';
 import { DuplicateEmailError } from '@/backend/auth/signup/dtos/DuplicateEmailError';
 import { ValidationError } from '@/backend/auth/signup/dtos/ValidationError';
 import { AuthError } from '@/backend/auth/signup/dtos/AuthError'; 
 import { SupabaseClient } from '@supabase/supabase-js';
-import { createClient } from '@/backend/common/infrastructure/supabase/server';
+import { createClient } from '@/backend/common/infrastructures/supabase/server';
 
 // 의존성 생성 팩토리 - 테스트하기 쉽고 이해하기 명확함
 async function createRegisterUseCase(): Promise<RegisterUseCase> {
