+++ backend/admin/spaces/usecases/PostSpaceUsecase.ts
@@ -1,6 +1,6 @@
-import { SpaceRepository } from "@/backend/common/domain/repository/SpaceRepository";
+import { SpaceRepository } from "@/backend/common/domains/repositories/SpaceRepository";
 import { PostSpaceQueryDto } from "../dtos/PostSpaceQueryDto";
-import { SaveRequest } from "@/backend/common/domain/repository/spaceRequest";
+import { SaveRequest } from "@/backend/common/domains/repositories/spaceRequest";
 
 export class PostSpaceUsecase {
   private repository: SpaceRepository;
