+++ backend/common/domains/entities/RoomItem.ts
