+++ backend/common/domains/entities/Supply.ts
