+++ backend/common/domains/repositories/SpaceRepository.ts
