+++ backend/common/domains/types/UserType.d.ts
