+++ backend/common/infrastructures/next-auth/next-auth.d.ts
