+++ backend/common/infrastructures/utils/CaseConvertUtils.ts
