+++ backend/user/reservations/usecases/DeleteUserRsvUsecase.ts
@@ -1,4 +1,4 @@
-import { RsvRepository } from "@/backend/common/domain/repository/RsvRepository";
+import { RsvRepository } from "@/backend/common/domains/repositories/RsvRepository";
 import { DeleteUserRsvDto } from "../dtos/DeleteUserRsvDto";
 
 export class DeleteUserRsvUsecase {
